Currently YAHC does not allow to override the socket options.  This provides the flexibilty to define the socket options as per the different needs of the system.
The socket options could be defined per request and could be used in
the request retries as well.

The socket options were set while building the socket and would be
reused for other requests using the same cached socket.  One can now
specify the socket options which can be set to the cached socket
before it is used for the connection.

However, the clients need to take care of passing socket options in
their requests if they do not want the previously set socket
options.

Note: Overriding socket options is not recommended with socket caching
as it could poison cached sockets with previously set socket options.